+++ lib/iso8601/date.rb
@@ -1,14 +1,14 @@
 module ISO8601
   ##
-  # A Date representation
+  # A Date representation.
   #
   # @example
-  #     d = Date.new('2014-05-28')
+  #     d = ISO8601::Date.new('2014-05-28')
   #     d.year  # => 2014
   #     d.month # => 5
   #
   # @example Week dates
-  #     d = Date.new('2014-W15-2')
+  #     d = ISO8601::Date.new('2014-W15-2')
   #     d.day   # => 27
   #     d.wday  # => 2
   #     d.week # => 15
@@ -17,7 +17,7 @@ module ISO8601
 
     def_delegators(:@date,
       :to_s, :to_time, :to_date, :to_datetime,
-      :year, :month, :day, :wday, :hash)
+      :year, :month, :day, :wday)
     ##
     # The original atoms
     attr_reader :atoms
@@ -60,6 +60,25 @@ module ISO8601
     # Converts self to an array of atoms.
     def to_a
       [year, month, day]
+    end
+    ##
+    # @param [#hash] contrast The contrast to compare against
+    #
+    # @return [Boolean]
+    def ==(contrast)
+      (hash == contrast.hash)
+    end
+    ##
+    # @param [#hash] contrast The contrast to compare against
+    #
+    # @return [Boolean]
+    def eql?(contrast)
+      (hash == contrast.hash)
+    end
+    ##
+    # @return [Fixnum]
+    def hash
+      [atoms, self.class].hash
     end
 
     private