+++ spec/iso8601/time_spec.rb
@@ -32,52 +32,51 @@ describe ISO8601::Time do
   context 'reduced patterns' do
     it "should parse correctly reduced times" do
       reduced_time = ISO8601::Time.new('T101112Z')
-      reduced_time.hour.should == 10
-      reduced_time.minute.should == 11
-      reduced_time.second.should == 12
+      expect(reduced_time.hour).to eq(10)
+      expect(reduced_time.minute).to eq(11)
+      expect(reduced_time.second).to eq(12)
     end
   end
 
   it "should return each atomic value" do
-    dt = ISO8601::Time.new('T12:02:01+04:00', ::Date.parse('2010-05-09'))
-    dt.hour.should == 12
-    dt.minute.should == 2
-    dt.second.should == 1
-    dt.zone.should == '+04:00'
+    t = ISO8601::Time.new('T12:02:01+04:00', ::Date.parse('2010-05-09'))
+    expect(t.hour).to eq(12)
+    expect(t.minute).to eq(2)
+    expect(t.second).to eq(1)
+    expect(t.zone).to eq('+04:00')
   end
 
   it "should respond to delegated casting methods" do
-    dt = ISO8601::Time.new('T10:09:08Z')
-    dt.should respond_to(:to_s, :to_time, :to_date, :to_datetime)
+    expect(ISO8601::Time.new('T10:09:08Z')).to respond_to(:to_s, :to_time, :to_date, :to_datetime)
   end
 
   describe '#+' do
     it "should return the result of the addition" do
-      (ISO8601::Time.new('T20:20:20+02:00') + 10).to_s.should == 'T20:20:30+02:00'
-      (ISO8601::Time.new('T20:20:20.11+02:00') + 10).to_s.should == 'T20:20:30.11+02:00'
+      expect((ISO8601::Time.new('T20:20:20+02:00') + 10).to_s).to eq('T20:20:30+02:00')
+      expect((ISO8601::Time.new('T20:20:20.11+02:00') + 10).to_s).to eq('T20:20:30.11+02:00')
     end
   end
 
   describe '#-' do
     it "should return the result of the substraction" do
-      (ISO8601::Time.new('T20:20:20+01:00') - 10).to_s.should == 'T20:20:10+01:00'
-      (ISO8601::Time.new('T20:20:20.11+02:00') - 10).to_s.should == 'T20:20:10.11+02:00'
+      expect((ISO8601::Time.new('T20:20:20+01:00') - 10).to_s).to eq('T20:20:10+01:00')
+      expect((ISO8601::Time.new('T20:20:20.11+02:00') - 10).to_s).to eq('T20:20:10.11+02:00')
     end
   end
 
   describe '#to_a' do
     it "should return an array of atoms" do
-      ISO8601::Time.new('T19:29:39Z').to_a.should == [19, 29, 39, '+00:00']
+      expect(ISO8601::Time.new('T19:29:39Z').to_a).to eq([19, 29, 39, '+00:00'])
     end
   end
 
   describe '#atoms' do
     it "should return an array of atoms" do
-      ISO8601::Time.new('T19:29:39+04:00').atoms.should == [19, 29, 39, '+04:00']
-      ISO8601::Time.new('T19:29:39Z').atoms.should == [19, 29, 39, 'Z']
-      ISO8601::Time.new('T19:29:39').atoms.should == [19, 29, 39]
-      ISO8601::Time.new('T19:29').atoms.should == [19, 29]
-      ISO8601::Time.new('T19').atoms.should == [19]
+      expect(ISO8601::Time.new('T19:29:39+04:00').atoms).to eq([19, 29, 39, '+04:00'])
+      expect(ISO8601::Time.new('T19:29:39Z').atoms).to eq([19, 29, 39, 'Z'])
+      expect(ISO8601::Time.new('T19:29:39').atoms).to eq([19, 29, 39])
+      expect(ISO8601::Time.new('T19:29').atoms).to eq([19, 29])
+      expect(ISO8601::Time.new('T19').atoms).to eq([19])
     end
   end
 
@@ -89,10 +88,9 @@ describe ISO8601::Time do
     end
     it "should return the same hash" do
       subject = ISO8601::Time.new('T20:11:10Z')
+      contrast = ISO8601::Time.new('T20:11:10Z')
 
-      expect(subject.to_time.hash).to eq(ISO8601::Time.new('T20:11:10Z').to_time.hash)
-      expect(subject.hash).to eq(ISO8601::Time.new('T20:11:10Z').hash)
+      expect(subject.hash).to eq(contrast.hash)
     end
   end
-
 end