+++ CHANGELOG.md
@@ -1,22 +1,32 @@
-## Changes for 0.7.0
+## 0.8.0
+
+* `DateTime` has hash identity by value.
+* `Time` has hash identity by value.
+* `Date` has hash identity by value.
+* `Duration` has hash identity by value.
+* `Atom` has hash identity by value.
+* `Atom#value` returns either an integer or a float.
+* `Atom#to_s` returns a valid ISO8601 subpattern.
+
+## 0.7.0
 
 * Add decimal fractions for any component in a duration.
 * Add a catch all `ISO8601::Errors::StandardError`.
 * Add support for comma (`,`) as a separator for duration decimal fractions.
 
-## Changes for 0.6.0
+## 0.6.0
 
 * Add `#hash` to `Duration`, `Date`, `Time` and `DateTime`.
 
-## Changes for 0.5.2
+## 0.5.2
 
 * Fix `DateTime` when handling empty strings.
 
-## Changes for 0.5.1
+## 0.5.1
 
 * Fix durations with sign.
 
-## Changes for 0.5
+## 0.5.0
 
 * Drop support for Ruby 1.8.7.
 * Add support for Rubinius 2.